Added scrolling behavior for primitive console optimized memset64 in assembly with stosq added alpha to the loading image (kinda) other small changes i dont rember :|

+++ Kernel/src/Headers/Interrupts.h
@@ -128,7 +128,7 @@ namespace Interrupts
 		};
 	};
 
-	void RegisterInterrupt(void* handlerAddress, byte inter, IdtType interruptType = IdtType::InterruptGate);
+	void RegisterInterrupt(void* handlerAddress, u16 inter, IdtType interruptType = IdtType::InterruptGate);
 
 	void LoadGIDT();
